backend: add repos and some controllers
need to think about handling login and registration on the same page

+++ lab4/src/main/kotlin/lab4/model/Result.kt
@@ -5,13 +5,13 @@ import javax.persistence.*
 @Entity
 @Table(name = "hit_result")
 data class Result(var x: Double = 0.0,
-                         var y: Double = 0.0,
-                         var r: Double = 0.0,
-                         var isHit: Boolean = false) {
+                  var y: Double = 0.0,
+                  var r: Double = 0.0,
+                  var isHit: Boolean = false) {
 
     @Id
     @GeneratedValue
-    var id: Long? = null
+    var id: Int? = null
 
     @ManyToOne(fetch = FetchType.EAGER)
     @JoinColumn(name = "user_id", referencedColumnName = "id", nullable = false)
